# meal-planner/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Healthy Recipe Finder API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Recipe {
  id: string;
  name: string;
  /** breakfast, lunch, dinner, or snack */
  category: string;
  /** veg, non-veg, vegan, jain, or kosher */
  mealType: string;
  ingredients: string[];
  allergens: string[];
  instructions: string;
  servings: string;
  /** @nullable */
  calories?: number | null;
  /** @nullable */
  protein?: string | null;
  /** @nullable */
  carbs?: string | null;
  /** @nullable */
  fat?: string | null;
  /** @nullable */
  prepTime?: string | null;
  /**
     * Percentage of available ingredients matched (0-100)
     * @nullable
     */
  matchScore?: number | null;
  /** @nullable */
  source?: string | null;
}

export interface RecipeListResponse {
  recipes: Recipe[];
  total: number;
}

export type RecipeStatsByCategory = {[key: string]: number};

export type RecipeStatsByMealType = {[key: string]: number};

export interface RecipeStats {
  totalRecipes: number;
  byCategory: RecipeStatsByCategory;
  byMealType: RecipeStatsByMealType;
}

export interface ErrorResponse {
  error: string;
}

export type ListRecipesParams = {
/**
 * Comma-separated list of available ingredients
 */
ingredients?: string;
/**
 * Comma-separated list of allergens to exclude
 */
allergies?: string;
/**
 * Filter by dietary type (veg, non-veg, vegan, jain, kosher)
 */
dietaryType?: string;
/**
 * Filter by meal category (breakfast, lunch, dinner, snack)
 */
category?: string;
/**
 * Sort field (name, calories, matchScore)
 */
sortBy?: string;
/**
 * Sort order (asc, desc)
 */
sortOrder?: string;
};