# meal-planner/lib/api-zod/src/generated/types/recipe.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Healthy Recipe Finder API
 * OpenAPI spec version: 0.1.0
 */

export interface Recipe {
  id: string;
  name: string;
  /** breakfast, lunch, dinner, or snack */
  category: string;
  /** veg, non-veg, vegan, jain, or kosher */
  mealType: string;
  ingredients: string[];
  allergens: string[];
  instructions: string;
  servings: string;
  /** @nullable */
  calories?: number | null;
  /** @nullable */
  protein?: string | null;
  /** @nullable */
  carbs?: string | null;
  /** @nullable */
  fat?: string | null;
  /** @nullable */
  prepTime?: string | null;
  /**
     * Percentage of available ingredients matched (0-100)
     * @nullable
     */
  matchScore?: number | null;
  /** @nullable */
  source?: string | null;
}