# meal-planner/lib/api-zod/src/generated/types/recipeListResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Healthy Recipe Finder API
 * OpenAPI spec version: 0.1.0
 */
import type { Recipe } from './recipe';

export interface RecipeListResponse {
  recipes: Recipe[];
  total: number;
}