# meal-planner/lib/api-zod/src/generated/types/recipeStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Healthy Recipe Finder API
 * OpenAPI spec version: 0.1.0
 */
import type { RecipeStatsByCategory } from './recipeStatsByCategory';
import type { RecipeStatsByMealType } from './recipeStatsByMealType';

export interface RecipeStats {
  totalRecipes: number;
  byCategory: RecipeStatsByCategory;
  byMealType: RecipeStatsByMealType;
}